# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: LiteShop API
  description: LiteShop 自动发卡系统 API。公共端点无需认证；管理端点需先登录获取会话 Cookie（角色要求：admin / operator
    / viewer）。支付回调路径可在后台配置（默认 /notify/bepusdt、/notify/hashpay）。
  version: 0.1.0
  contact:
    name: LiteShop
    url: https://github.com/mhan24/liteshop
servers:
- url: /
...
security: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: shop_session
      description: 登录后由服务端设置的会话 Cookie
  schemas:
...
    LoginRequest:
      type: object
      required:
      - username
      - password
      properties:
        username:
          type: string
        password:
          type: string
        otp:
          type: string
          description: TOTP 验证码（已启用双因素时）
    LoginResponse:
      type: object
      properties:
        ok:
          type: boolean
        totp_required:
          type: boolean
        token:
          type: string
          description: TOTP 待验证令牌
...
paths:
...
  /api/v1/admin/login:
    post:
      tags:
      - admin
      summary: 管理员登录
      description: 成功设置会话 Cookie；已启用 TOTP 且未提供 otp 时返回 totp_required=true 与待验证 token。连续失败
        5 次锁定 10 分钟。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '403':
          description: 凭据错误 / 已锁定
...